manager/requirements.txt
aiofiles
fastapi
fastapi_login
hypercorn
natsort
Pillow
psutil
pydantic
PyJWT
pulsectl
pymediainfo
python_crontab
python-multipart
qrcode
rpyc
ruamel.yaml
uvloop
validators
jinja2
Werkzeug